connexion-101/requirements-api.txt
#
# Further dependencies for API training
#
connexion[swagger-ui]
connexion
openapi-resolver
pyyaml
flask_testing
flask
pytz
